refactor: changes text on about section, re-organize navbar links
