+++ src/components/Templates/AboutSection.tsx
@@ -38,14 +38,9 @@ const AboutSection = ({ id }: IAppSection) => (
               </p>
               <p className={'mt-3 font-light text-current'}>
                 HDX rewards are generated by bonding funds to the protocol to
-                provide it further liquidity and by staking to help secure the
-                network. HDX will additionally serve as the meta-token of our
-                Layer 3 DEX, being the end recipient of swaps from trading fees
-                and DeFi services.” to “HDX rewards are generated by bonding
-                funds to the protocol to provide it with further liquidity. HDX
-                will additionally serve as the meta-token of our Layer 3 DEX,
-                being the end recipient of swaps from trading fees and DeFi
-                services.
+                provide it with further liquidity. HDX will additionally serve
+                as the meta-token of our Layer 3 DEX, being the end recipient of
+                swaps from trading fees and DeFi services.
               </p>
             </div>
           </>
